rename Ze main package and zeclasses
