+++ zeapp/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.ZeBadgeApp"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ZeMainActivity"
             android:exported="true"
             android:theme="@style/Theme.ZeBadgeApp">
             <intent-filter>
