+++ zeapp/app/src/main/java/de/berlindroid/zeapp/ZeConfig.kt
