+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zeui/ZeNameEditorDialog.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.window.DialogProperties
 import de.berlindroid.zeapp.zebits.composableToBitmap
 import de.berlindroid.zeapp.zebits.isBinary
 import de.berlindroid.zeapp.zeui.zepages.NamePage
-import de.berlindroid.zeapp.vm.ZeBadgeViewModel.Configuration
+import de.berlindroid.zeapp.zevm.ZeBadgeViewModel.Configuration
 
 
 /**
